- dump memory feature - small fixes - changed unordered_map to hash_map again

+++ Scylla/ImportRebuild.cpp
@@ -200,8 +200,8 @@ bool ImportRebuild::saveNewFile(const WCHAR * filepath)
 
 	if(pOverlay)
 	{
-		ProcessAccessHelp::writeMemoryToFile(hFile, fileOffset, sizeOfOverlay, pOverlay);
-		fileOffset += sizeOfOverlay;
+		ProcessAccessHelp::writeMemoryToFile(hFile, fileOffset, (DWORD)sizeOfOverlay, pOverlay);
+		fileOffset += (DWORD)sizeOfOverlay;
 	}
 
 	CloseHandle(hFile);
